Reorganize and rename folders and files
