+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
-client/dist/
-config/
+src/config/
+src/dist/
 stubs/
