--- client/styles.css
@@ -1,3 +1,0 @@
-body {
-	background-color: #555555;
-}
