+++ package.json
@@ -14,7 +14,6 @@
     "react": "^15.5.3",
     "react-bootstrap": "^0.30.7",
     "react-dom": "^15.5.3",
-    "react-mapbox-gl": "^1.8.0",
     "react-redux": "^5.0.3",
     "react-router": "^3.0.2",
     "redux": "^3.6.0",
@@ -36,7 +35,7 @@
     "webpack": "^2.2.1"
   },
   "scripts": {
-    "start": "nodemon server.js & webpack --watch --progress"
+    "start": "nodemon app.js & webpack --watch --progress"
   },
   "repository": {
     "type": "git",
