+++ src/components/map/map.scss
@@ -1,5 +1,5 @@
 #map-display {
-	width: 100%;
-	height: 100%;
+	width: 70%;
+	height: 70%;
 	margin: auto;
 }
