+++ src/index.scss
@@ -1,0 +1,29 @@
+$font-stack: 'Roboto', Helvetica, sans-serif;
+$primary-color: #999;
+
+html, body, form {
+  font: 100% $font-stack;
+	margin: 0;
+	padding: 0;
+}
+
+body {
+  background-color: $primary-color;
+}
+
+a {
+	color: #039;
+}
+
+a:hover {
+	color: #036;
+}
+
+p {
+	padding-bottom: 1.3em;
+}
+
+#header, #footer {
+	width: 100%;
+	margin: auto;
+}
