+++ src/reducers/mapTime.js
